Extract Rule and RuleProperty utils from model and fix tests
