+++ src/js/models/Rule.js
@@ -11,7 +11,6 @@
 import { Record, Map } from 'immutable';
 import { RuleDefinitionFactory } from './RuleDefinition.js';
 import { RulePropertyFactory } from './RuleProperty';
-import { RulePropertyUtils } from './RuleProperty';
 import type { RuleDefinition } from './RuleDefinition.js';
 import type { RuleProperty } from './RuleProperty';
 
@@ -61,16 +60,5 @@ export const RuleFactory = (values: $Shape<RuleRecord>): Rule => {
 
   return rule;
 };
-
-export class RuleUtils {
-  static isValid(rule: Rule) {
-    if (rule.selector == '' || rule.selector == null) {
-      return false;
-    }
-    return rule.properties
-      .filter(property => property.definition.required)
-      .every(property => RulePropertyUtils.isValid(property));
-  }
-}
 
 export type Rule = RecordOf<RuleRecord> & RuleFactory;
